pyproject.toml
[tool.poetry]
name = "tapioca-arbache"
version = "0.1.0"
description = "Clients for Arbache's APIs"
authors = ["Leonardo Jose N Silva"]

[tool.poetry.dependencies]
python = "^3.9"
tapioca-wrapper = "^2.0.1"
requests-oauthlib = "^1.3.0"
responses = "^0.13.3"
python-status = "^1.0.1"

[tool.poetry.dev-dependencies]
ipython = "^7.25.0"
pytest = "^6.2.4"
twine = "^3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
